Add test run Int24bit native c++tests

+++ PasswordsAPI.Tests/PasswordsAPI.Tests.cs
@@ -6,13 +6,46 @@ using Xunit;
 
 namespace Passwords.API.Tests
 {
+    public class BuildConfig
+    {
+        public readonly string Architecture;
+        public readonly string Configuration;
+        public BuildConfig()
+        {
+            Architecture = Consola.StdStream.MachineArch().Contains("64bit")
+                         ? "x64" : "x86";
+#if DEBUG
+            Configuration = "Debug";
+#else
+                Configuration = "Release";
+#endif
+        }
+    }
+
     public class ExternalLibraryTestRuns
     {
-        private static Tuple<int, string> executeLibraryTest(string path,string dllname)
+
+        private BuildConfig _buildConfig;
+        private BuildConfig BuildConfig
         {
-            ProcessStartInfo startinfo = new ProcessStartInfo(
-                "C:\\Program Files\\dotnet\\dotnet.exe",
-                   $"{path}\\{dllname}.dll --verbose" );
+            get { return _buildConfig ?? (_buildConfig = new BuildConfig()); }
+        }
+
+
+        private static Tuple<int, string> executeLibraryTest(string path, string file)
+        {
+            ProcessStartInfo startinfo;
+            bool ConsolaTest;
+            if (file.EndsWith(".exe")) {
+                ConsolaTest = false;
+                startinfo = new ProcessStartInfo($"{path}\\{file}");
+            } else if (file.EndsWith(".dll")) {
+                ConsolaTest = true;
+                startinfo = new ProcessStartInfo(
+                    "C:\\Program Files\\dotnet\\dotnet.exe",
+                    $"{path}\\{file} --verbose" );
+            } else 
+                return new Tuple<int,string>( int.MaxValue, file+" is not executable" );
 
             startinfo.RedirectStandardOutput = true;
             startinfo.RedirectStandardError = true;
@@ -22,29 +55,38 @@ namespace Passwords.API.Tests
             testrun.EnableRaisingEvents = true;
             bool ok = false;
             System.Text.StringBuilder testoutput = new System.Text.StringBuilder();
-            if (testrun.Start()) {
-                testrun.WaitForExit(3000);
-                ok = true;
-            } else {
-                testoutput.Append("Unknown Status running test");
-            }
 
-            if (ok) {
-                FileInfo outfile =
-                    new FileInfo($"{path}\\dotnet_Err.log");
-                if (outfile.Exists) {
-                    StreamReader f = outfile.OpenText();
-                    testoutput.Append(f.ReadToEnd()).Append("\n");
-                    f.Close();
-                    outfile.Delete();
-                    outfile = new FileInfo(
-                        $"{path}\\dotnet_Out.log");
-                    f = outfile.OpenText();
-                    testoutput.Append(f.ReadToEnd());
-                    f.Close();
-                    outfile.Delete();
+            if (ConsolaTest) {
+                if (testrun.Start()) {
+                    ok = testrun.WaitForExit(300000);
                 } else {
-                    testoutput.Append("Unknown Status reading output");
+                    testoutput.Append("Unknown Status running test");
+                }
+                if (ok) {
+                    FileInfo outfile =
+                        new FileInfo($"{path}\\dotnet_Err.log");
+                    if (outfile.Exists) {
+                        StreamReader f = outfile.OpenText();
+                        testoutput.Append(f.ReadToEnd()).Append("\n");
+                        f.Close();
+                        outfile.Delete();
+                        outfile = new FileInfo($"{path}\\dotnet_Out.log");
+                        f = outfile.OpenText();
+                        testoutput.Append(f.ReadToEnd());
+                        f.Close();
+                        outfile.Delete();
+                    } else {
+                        testoutput.Append("Unknown Status reading output");
+                    }
+                }
+            } else {
+                if (testrun.Start()) {
+                    if (ok = testrun.WaitForExit(30000)) {
+                        testoutput.Append(testrun.StandardError.ReadToEnd());
+                        testoutput.Append(testrun.StandardOutput.ReadToEnd());
+                    } else testoutput.Append("test hanging crashed... ");
+                } else {
+                    testoutput.Append("Unknown status reading output");
                 }
             }
             Tuple<int, string> result = new Tuple<int,string>(testrun.ExitCode, testoutput.ToString());
@@ -53,16 +95,31 @@ namespace Passwords.API.Tests
         }
 
         [Fact]
-        public void RunInt24TypesTest()
+        public void RunInt24NativeTypesTest()
         {
-            Tuple<int, string> result = executeLibraryTest("C:\\WORKSPACE\\PROJECTS\\Int24Types\\bin\\core5\\test\\x64\\Debug\\net5.0", "test_int24_dotnet_dll");
-            Assert.True( result.Item1 == 0, result.Item2 );
+            string path = "C:\\WORKSPACE\\PROJECTS\\Int24Types\\bin\\native\\" +
+                        $"{BuildConfig.Architecture}\\{BuildConfig.Configuration}";
+
+            Tuple<int, string> result = executeLibraryTest( path, "test_int24_native_cpp.exe" );
+            Assert.True(result.Item1 == 0, result.Item1.ToString() + "failures... " + result.Item2);
+        }
+
+        [Fact]
+        public void RunInt24DotnetTypesTest()
+        {
+            string path = "C:\\WORKSPACE\\PROJECTS\\Int24Types\\bin\\core5\\test\\" +
+                $"{BuildConfig.Architecture}\\{BuildConfig.Configuration}\\net5.0";
+
+            Tuple<int,string> result = executeLibraryTest( path, "test_int24_dotnet_dll.dll" );
+            Assert.True(result.Item1 == 0, result.Item2);
         }
 
         [Fact]
         public void RunYpsCryptTests()
         {
-            Tuple<int, string> result = executeLibraryTest("C:\\WORKSPACE\\PROJECTS\\PasswordsManager\\PasswordsAPI\\bin\\x64\\Debug\\net5.0", "YpsCryps");
+            string path = "C:\\WORKSPACE\\PROJECTS\\PasswordsManager\\PasswordsAPI\\bin"
+                + $"\\{BuildConfig.Architecture}\\{BuildConfig.Configuration}\\net5.0";
+            Tuple<int, string> result = executeLibraryTest( path, "YpsCryps.dll" );
             Assert.True( result.Item1 == 0, result.Item2 );
         }
     }
